chore: merge swapperApi into swappersApi
