+++ src/features/defi/providers/thorchain-savers/components/ThorchainSaversManager/Deposit/components/Confirm.tsx
@@ -58,7 +58,7 @@ import {
 import { fromThorBaseUnit, getThorchainFromAddress, toThorBaseUnit } from 'lib/utils/thorchain'
 import { BASE_BPS_POINTS } from 'lib/utils/thorchain/constants'
 import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
-import { getIsTradingActiveApi } from 'state/apis/swapper/getIsTradingActiveApi'
+import { swappersApi } from 'state/apis/swappers/swappersApi'
 import {
   getMaybeThorchainSaversDepositQuote,
   getThorchainSaversPosition,
@@ -561,7 +561,7 @@ export const Confirm: React.FC<ConfirmProps> = ({ accountId, onNext }) => {
 
       contextDispatch({ type: ThorchainSaversDepositActionType.SET_LOADING, payload: true })
 
-      const { getIsTradingActive } = getIsTradingActiveApi.endpoints
+      const { getIsTradingActive } = swappersApi.endpoints
       const { data: isTradingActive } = await appDispatch(
         getIsTradingActive.initiate({
           assetId,
