+++ src/features/defi/providers/thorchain-savers/components/ThorchainSaversManager/Overview/ThorchainSaversOverview.tsx
@@ -36,8 +36,8 @@ import { Text } from 'components/Text'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit, toBaseUnit } from 'lib/math'
-import { getIsTradingActiveApi } from 'state/apis/swapper/getIsTradingActiveApi'
-import { selectSwapperApiTradingActivePending } from 'state/apis/swapper/selectors'
+import { selectSwapperApiTradingActivePending } from 'state/apis/swappers/selectors'
+import { swappersApi } from 'state/apis/swappers/swappersApi'
 import type { ThorchainSaversStakingSpecificMetadata } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/types'
 import {
   getMaybeThorchainSaversDepositQuote,
@@ -134,7 +134,7 @@ export const ThorchainSaversOverview: React.FC<OverviewProps> = ({
 
   useEffect(() => {
     ;(async () => {
-      const { getIsTradingActive } = getIsTradingActiveApi.endpoints
+      const { getIsTradingActive } = swappersApi.endpoints
       const { data: isTradingActive } = await appDispatch(
         getIsTradingActive.initiate({
           assetId,
