--- src/state/apis/swapper/selectors.ts
@@ -1,7 +1,0 @@
-import { QueryStatus } from '@reduxjs/toolkit/query'
-import type { ReduxState } from 'state/reducer'
-
-export const selectSwapperApiTradingActivePending = (state: ReduxState) =>
-  Object.values(state.swapperApi.queries).some(
-    query => query?.endpointName === 'getIsTradingActive' && query?.status === QueryStatus.pending,
-  )
