--- src/state/apis/swapper/swapperApi.ts
@@ -1,8 +1,0 @@
-import { createApi } from '@reduxjs/toolkit/query/react'
-import { BASE_RTK_CREATE_API_CONFIG } from 'state/apis/const'
-
-export const swapperApi = createApi({
-  ...BASE_RTK_CREATE_API_CONFIG,
-  reducerPath: 'swapperApi',
-  endpoints: () => ({}),
-})
